refactor(api): move entity-version write-back helper to lib (keep vertical-slice metric flat)
edit_workspace_document's DOCX write-back is genuinely shared infrastructure,
not chat-specific: relocate createEntityVersionFromBuffer,
loadEntityVersionDocxBuffer, and validateDocxBuffer out of the entities
handler domain into apps/api/src/lib/entity-versions/ so the chat tool no
longer reaches across handler domains for them.

As a side effect, folio-collab/finalize.ts's own pre-existing reach into
handlers/entities for validateDocxBuffer also drops, so cross-handler-imports
lands one below the pre-edit-tool baseline (252 vs. 253); baseline rewritten
via `ratchet.ts --write` to match the live count.
