+++ apps/api/src/mcp/template-tools.ts
@@ -12,7 +12,6 @@ import {
 } from "@/api/handlers/docx/ai-field-generator";
 import type { FieldMeta, FieldPart } from "@/api/handlers/docx/types";
 import { INPUT_TYPES, isFieldMeta } from "@/api/handlers/docx/types";
-import { validateDocxBuffer } from "@/api/handlers/entities/validate-docx-buffer";
 import { configureTemplateFields } from "@/api/handlers/templates/configure-template-fields-service";
 import { createStoredTemplate } from "@/api/handlers/templates/create-template-service";
 import { recordTemplateFill } from "@/api/handlers/templates/record-use";
@@ -26,6 +25,7 @@ import { loadOrgAIConfig } from "@/api/lib/ai-config-loader";
 import { captureError } from "@/api/lib/analytics/capture";
 import { createTanStackAIAnalyticsCallbacks } from "@/api/lib/analytics/tanstack-ai";
 import { assertUsageAvailableForHandler } from "@/api/lib/api-handlers";
+import { validateDocxBuffer } from "@/api/lib/entity-versions/validate-docx-buffer";
 import { FILE_SIZE_LIMIT_BYTES, LIMITS } from "@/api/lib/limits";
 import {
   createCursorPage,
